+++ packages/neutron-sdk/src/interchain_txs/helpers.rs
@@ -1,5 +1,9 @@
+use crate::bindings::msg::ChannelOrdering;
+use crate::NeutronResult;
 use cosmos_sdk_proto::traits::Message;
-use cosmwasm_std::{StdError, StdResult};
+use cosmwasm_std::{Addr, CosmosMsg, StdError, StdResult};
+use neutron_std::types::cosmos::base::v1beta1::Coin;
+use neutron_std::types::neutron::interchaintxs::v1::MsgRegisterInterchainAccount;
 
 /// Decodes protobuf any item into T structure
 pub fn decode_message_response<T: Message + Default>(item: &Vec<u8>) -> StdResult<T> {
@@ -20,4 +24,21 @@ pub fn get_port_id<R: AsRef<str>>(contract_address: R, interchain_account_id: R)
         + contract_address.as_ref()
         + ICA_OWNER_DELIMITER
         + interchain_account_id.as_ref()
+}
+
+pub fn register_interchain_account(
+    contract: Addr,
+    connection_id: String,
+    interchain_account_id: String,
+    register_fee: Vec<Coin>,
+    ordering: Option<ChannelOrdering>,
+) -> NeutronResult<CosmosMsg> {
+    Ok(MsgRegisterInterchainAccount {
+        from_address: contract.to_string(),
+        connection_id,
+        interchain_account_id,
+        register_fee,
+        ordering: ordering.unwrap_or(ChannelOrdering::OrderOrdered).into(),
+    }
+    .into())
 }
